+++ lib/screens/alarm_ring_screen.dart
@@ -2,6 +2,7 @@ import 'package:flutter/material.dart';
 import 'package:provider/provider.dart';
 import '../models/alarm_model.dart';
 import '../services/alarm_service.dart';
+import '../services/alarm_settings.dart';
 import '../widgets/lock_screen.dart';
 
 class AlarmRingScreen extends StatefulWidget {
@@ -36,12 +37,16 @@ class _AlarmRingScreenState extends State<AlarmRingScreen> {
   Widget build(BuildContext context) {
     final now = TimeOfDay.now();
     final timeStr = now.format(context);
+    final settings = context.watch<AlarmSettings>();
 
     return LockScreen(
       mode: LockScreenMode.alarm,
       title: widget.alarm.label.isNotEmpty ? widget.alarm.label : 'Alarm',
       timeDisplay: timeStr,
       showSnooze: true,
+      autoDismissMinutes: settings.autoDismissMinutes,
+      vibrateEnabled: settings.vibrateOnAlarm,
+      volume: settings.volume,
       onStop: _dismiss,
       onSnooze: _snooze,
     );